Fix: Could not find required intl object. <IntlProvider> needs to exist in the component ancestry.
